requirements.txt
openai
selenium
requests
openai-whisper
pydub
audioop-lts